v5.6.8

+++ includes/class-wcj-payment-gateways.php
@@ -2,7 +2,7 @@
 /**
  * Booster for WooCommerce - Module - Custom Gateways
  *
- * @version 5.6.7
+ * @version 5.6.8
  * @author  Pluggabl LLC.
  * @package Booster_For_WooCommerce/includes
  */
@@ -20,7 +20,7 @@ if ( ! class_exists( 'WCJ_Payment_Gateways' ) ) :
 		/**
 		 * Constructor.
 		 *
-		 * @version 5.6.2
+		 * @version 5.6.8
 		 */
 		public function __construct() {
 
@@ -32,7 +32,7 @@ if ( ! class_exists( 'WCJ_Payment_Gateways' ) ) :
 			parent::__construct();
 
 			if ( $this->is_enabled() ) {
-				include_once 'gateways/class-wc-gateway-wcj-custom.php';
+				include_once 'gateways/class-wc-gateway-wcj-custom-template.php';
 				add_action( 'woocommerce_after_checkout_validation', array( $this, 'check_required_wcj_input_fields' ), PHP_INT_MAX, 2 );
 				add_action( 'woocommerce_checkout_update_order_meta', array( $this, 'update_custom_payment_gateways_fields_order_meta' ), PHP_INT_MAX, 2 );
 				add_action( 'add_meta_boxes', array( $this, 'add_custom_payment_gateways_fields_admin_order_meta_box' ) );
@@ -43,14 +43,15 @@ if ( ! class_exists( 'WCJ_Payment_Gateways' ) ) :
 		/**
 		 * Maybe_delete_payment_gateway_input_fields.
 		 *
-		 * @version 5.6.2
+		 * @version 5.6.8
 		 * @since   3.3.0
 		 */
 		public function maybe_delete_payment_gateway_input_fields() {
-			if ( isset( $_GET['wcj_delete_payment_gateway_input_fields'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification
-				$order_id = sanitize_text_field( wp_unslash( $_GET['wcj_delete_payment_gateway_input_fields'] ) ); // phpcs:ignore WordPress.Security.NonceVerification
+			$wpnonce = isset( $_GET['_wpnonce'] ) ? wp_verify_nonce( sanitize_key( $_GET['_wpnonce'] ), 'wcj_delete_payment_gateway_input_fields' ) : false;
+			if ( $wpnonce && isset( $_GET['wcj_delete_payment_gateway_input_fields'] ) ) {
+				$order_id = sanitize_text_field( wp_unslash( $_GET['wcj_delete_payment_gateway_input_fields'] ) );
 				delete_post_meta( $order_id, '_wcj_custom_payment_gateway_input_fields' );
-				wp_safe_redirect( esc_url( remove_query_arg( 'wcj_delete_payment_gateway_input_fields' ) ) );
+				wp_safe_redirect( esc_url_raw( remove_query_arg( array( 'wcj_delete_payment_gateway_input_fields', '_wpnonce' ) ) ) );
 				exit;
 			}
 		}
@@ -111,7 +112,7 @@ if ( ! class_exists( 'WCJ_Payment_Gateways' ) ) :
 		/**
 		 * Create_custom_payment_gateways_fields_admin_order_meta_box.
 		 *
-		 * @version 5.5.9
+		 * @version 5.6.8
 		 * @since   2.5.2
 		 */
 		public function create_custom_payment_gateways_fields_admin_order_meta_box() {
@@ -130,7 +131,7 @@ if ( ! class_exists( 'WCJ_Payment_Gateways' ) ) :
 				)
 			);
 			if ( 'yes' === wcj_get_option( 'wcj_custom_payment_gateways_input_fields_delete_button', 'no' ) ) {
-				$html .= '<p><a style="color:#a00;" href="' . esc_url( add_query_arg( 'wcj_delete_payment_gateway_input_fields', $order_id ) ) . '"' . wcj_get_js_confirmation() . '>' .
+				$html .= '<p><a style="color:#a00;" href="' . esc_url( wp_nonce_url( add_query_arg( 'wcj_delete_payment_gateway_input_fields', $order_id ), 'wcj_delete_payment_gateway_input_fields' ) ) . '"' . wcj_get_js_confirmation() . '>' .
 				__( 'Delete', 'woocommerce-jetpack' ) . '</a></p>';
 			}
 			echo wp_kses_post( $html );
